requirements.txt
Flask>=3.0,<4
flask-cors>=4.0,<5
pymongo[srv]>=4.6,<5
requests>=2.31,<3
cryptography>=41,<43
numpy>=1.26,<3
pandas>=2.2,<3
pytrends>=4.9.2,<5
python-dotenv>=1.0,<2
pyppeteer>=2.0,<3

# dev/test
pytest>=8,<9
responses>=0.25,<1
mongomock>=4.1,<5
freezegun>=1.5,<2
